Move InMemoryFeedStore out of tests to production code.
